+++ dot-strive/src/app/trainees/[trainee_id]/(private)/trainings/dates/[year]/[month]/[date]/page.tsx
@@ -9,10 +9,13 @@ import { css } from "styled-system/css";
 import { stack } from "styled-system/patterns";
 
 import { DailyTrainingList } from "./_components/daily-training-list";
-import { WeeklyTrainingCalendar } from "./_components/weekly-training-calendar";
+import { getWeeklyTrainings } from "./_repository/get-weekly-trainings";
 
+import type { TraineeId } from "@/app/_schemas/trainee";
+import type { UTCDateString } from "@/app/_schemas/utc-date-string";
 import type { NextPage } from "@/app/_types/page";
 import type { Route } from "next";
+import type { FC } from "react";
 
 const Page: NextPage = async (props) => {
   const traineeIdParam = props.params?.["trainee_id"];
@@ -61,7 +64,7 @@ const Page: NextPage = async (props) => {
             />
           }
         >
-          <WeeklyTrainingCalendar traineeId={traineeId} selected={selected} />
+          <FetchWeeklyTrainings traineeId={traineeId} selected={selected} />
         </Suspense>
       </div>
       <Suspense fallback={<p>トレーニングデータを取得しています</p>}>
@@ -80,3 +83,26 @@ const Page: NextPage = async (props) => {
   );
 };
 export default Page;
+
+type Props = {
+  traineeId: TraineeId;
+  selected: UTCDateString;
+};
+const FetchWeeklyTrainings: FC<Props> = async (props) => {
+  const result = await getWeeklyTrainings({
+    traineeId: props.traineeId,
+    date: props.selected,
+  });
+  if (result.isErr()) {
+    return <p>トレーニングデータの取得に失敗しました</p>;
+  }
+  const trainings = result.value;
+
+  return (
+    <TrainingCalendarWeek
+      traineeId={props.traineeId}
+      selected={props.selected}
+      trainings={trainings}
+    />
+  );
+};
