+++ dot-strive/src/app/trainees/[trainee_id]/(private)/trainings/dates/[year]/[month]/page.tsx
@@ -4,19 +4,31 @@ import { redirect } from "next/navigation";
 import { Suspense } from "react";
 
 import { Loading } from "@/app/_components/loading";
+import { TrainingCalendarMonth } from "@/app/_components/training-calendar-month";
+import { utcDateStringSchema } from "@/app/_schemas/utc-date-string";
 import { css } from "styled-system/css";
 import { stack } from "styled-system/patterns";
 
-import { MonthlyTrainingList } from "./_components/monthly-training-list";
+import { getMonthlyTrainings } from "./_repositories/get-monthly-trainings";
+import { getTraineeBySession } from "../../../../_repositories/get-trainee-by-session";
+import { TrainingDetailView } from "../../../_components/training-detail";
 
+import type { TraineeId } from "@/app/_schemas/trainee";
 import type { NextPage } from "@/app/_types/page";
 import type { Route } from "next";
+import type { FC } from "react";
 
-const Page: NextPage = (props) => {
-  const traineeId = props.params?.["trainee_id"];
-  if (!traineeId) {
+const Page: NextPage = async (props) => {
+  const traineeIdParam = props.params?.["trainee_id"];
+  const getTraineeResult = await getTraineeBySession();
+  if (
+    getTraineeResult.isErr() ||
+    getTraineeResult.value.id !== traineeIdParam
+  ) {
     redirect("/" satisfies Route);
   }
+  const traineeId = getTraineeResult.value.id;
+
   const year = props.params?.["year"];
   const month = props.params?.["month"];
   if (!year || !month || isNaN(Date.parse(`${year}-${month}`))) {
@@ -52,9 +64,64 @@ const Page: NextPage = (props) => {
       <Suspense
         fallback={<Loading description="トレーニングデータを取得しています" />}
       >
-        <MonthlyTrainingList traineeId={traineeId} year={year} month={month} />
+        <FetchMonthlyTrainings
+          traineeId={traineeId}
+          year={year}
+          month={month}
+        />
       </Suspense>
     </section>
   );
 };
 export default Page;
+
+type Props = {
+  traineeId: TraineeId;
+  year: string;
+  month: string;
+};
+export const FetchMonthlyTrainings: FC<Props> = async (props) => {
+  const result = await getMonthlyTrainings({
+    traineeId: props.traineeId,
+    year: props.year,
+    month: props.month,
+  });
+
+  if (result.isErr()) {
+    return <p>トレーニングデータの取得に失敗しました</p>;
+  }
+  const trainings = result.value;
+
+  return (
+    <div className={stack({ direction: "column" })}>
+      <TrainingCalendarMonth
+        traineeId={props.traineeId}
+        trainings={trainings}
+        selected={utcDateStringSchema.parse(
+          new Date(`${props.year}-${props.month}-1`).toISOString()
+        )}
+      />
+      <ul className={stack({ direction: "column", gap: 12, p: 4 })}>
+        {trainings
+          .sort(
+            (a, b) => new Date(b.date).valueOf() - new Date(a.date).valueOf()
+          )
+          .map((training) => {
+            const styles = css({
+              border: "1px solid",
+            });
+
+            return (
+              <li key={training.id} className={styles}>
+                <Link
+                  href={`/trainees/${props.traineeId}/trainings/${training.id}`}
+                >
+                  <TrainingDetailView training={training} />
+                </Link>
+              </li>
+            );
+          })}
+      </ul>
+    </div>
+  );
+};
